refactor: move remaining OTA code into the corresponding mixin
