+++ packages/zwave-js/src/lib/node/mixins/70_FirmwareUpdate.ts
@@ -1,5 +1,6 @@
 import {
 	FirmwareDownloadStatus,
+	type FirmwareUpdateCapabilities,
 	type FirmwareUpdateMetaData,
 	FirmwareUpdateMetaDataCC,
 	FirmwareUpdateMetaDataCCGet,
@@ -9,6 +10,7 @@ import {
 	type FirmwareUpdateMetaDataCCRequestGet,
 	FirmwareUpdateMetaDataCCRequestReport,
 	FirmwareUpdateMetaDataCCStatusReport,
+	FirmwareUpdateMetaDataCCValues,
 	type FirmwareUpdateOptions,
 	type FirmwareUpdateProgress,
 	FirmwareUpdateRequestStatus,
@@ -37,6 +39,7 @@ import {
 	createDeferredPromise,
 } from "alcalzone-shared/deferred-promise";
 import { roundTo } from "alcalzone-shared/math";
+import { isArray } from "alcalzone-shared/typeguards";
 import {
 	type Task,
 	type TaskBuilder,
@@ -86,11 +89,91 @@ export interface NodeFirmwareUpdate {
 	 * Returns whether a firmware update is in progress for this node.
 	 */
 	isFirmwareUpdateInProgress(): boolean;
+
+	/**
+	 * Retrieves the firmware update capabilities of a node to decide which options to offer a user prior to the update.
+	 * This communicates with the node to retrieve fresh information.
+	 */
+	getFirmwareUpdateCapabilities(): Promise<FirmwareUpdateCapabilities>;
+
+	/**
+	 * Retrieves the firmware update capabilities of a node to decide which options to offer a user prior to the update.
+	 * This method uses cached information from the most recent interview.
+	 */
+	getFirmwareUpdateCapabilitiesCached(): FirmwareUpdateCapabilities;
 }
 
 export abstract class FirmwareUpdateMixin extends SchedulePollMixin
 	implements NodeFirmwareUpdate
 {
+	public async getFirmwareUpdateCapabilities(): Promise<
+		FirmwareUpdateCapabilities
+	> {
+		const api = this.commandClasses["Firmware Update Meta Data"];
+		const meta = await api.getMetaData();
+		if (!meta) {
+			throw new ZWaveError(
+				`Failed to request firmware update capabilities: The node did not respond in time!`,
+				ZWaveErrorCodes.Controller_NodeTimeout,
+			);
+		} else if (!meta.firmwareUpgradable) {
+			return {
+				firmwareUpgradable: false,
+			};
+		}
+
+		return {
+			firmwareUpgradable: true,
+			// TODO: Targets are not the list of IDs - maybe expose the IDs as well?
+			firmwareTargets: new Array(1 + meta.additionalFirmwareIDs.length)
+				.fill(0).map((_, i) => i),
+			continuesToFunction: meta.continuesToFunction,
+			supportsActivation: meta.supportsActivation,
+			supportsResuming: meta.supportsResuming,
+			supportsNonSecureTransfer: meta.supportsNonSecureTransfer,
+		};
+	}
+
+	public getFirmwareUpdateCapabilitiesCached(): FirmwareUpdateCapabilities {
+		const firmwareUpgradable = this.getValue<boolean>(
+			FirmwareUpdateMetaDataCCValues.firmwareUpgradable.id,
+		);
+		const supportsActivation = this.getValue<boolean>(
+			FirmwareUpdateMetaDataCCValues.supportsActivation.id,
+		);
+		const continuesToFunction = this.getValue<boolean>(
+			FirmwareUpdateMetaDataCCValues.continuesToFunction.id,
+		);
+		const additionalFirmwareIDs = this.getValue<number[]>(
+			FirmwareUpdateMetaDataCCValues.additionalFirmwareIDs.id,
+		);
+		const supportsResuming = this.getValue<boolean>(
+			FirmwareUpdateMetaDataCCValues.supportsResuming.id,
+		);
+		const supportsNonSecureTransfer = this.getValue<boolean>(
+			FirmwareUpdateMetaDataCCValues.supportsNonSecureTransfer.id,
+		);
+
+		// Ensure all information was queried
+		if (
+			!firmwareUpgradable
+			|| !isArray(additionalFirmwareIDs)
+		) {
+			return { firmwareUpgradable: false };
+		}
+
+		return {
+			firmwareUpgradable: true,
+			// TODO: Targets are not the list of IDs - maybe expose the IDs as well?
+			firmwareTargets: new Array(1 + additionalFirmwareIDs.length).fill(0)
+				.map((_, i) => i),
+			continuesToFunction,
+			supportsActivation,
+			supportsResuming,
+			supportsNonSecureTransfer,
+		};
+	}
+
 	private _abortFirmwareUpdate: (() => Promise<void>) | undefined;
 	public async abortFirmwareUpdate(): Promise<void> {
 		if (!this._abortFirmwareUpdate) return;
